feat(sessions): added goToSession button
